+++ R/create_feedback.R
@@ -4,6 +4,6 @@
 #'
 #' @export
 create_feedback = function(){
-  feedback = system.file("feedback/feedback.Rmd",package="jrPresentation")
+  feedback = system.file("feedback/feedback.Rmd", package = "jrPresentation")
   file.copy(feedback, to = "./feedback.Rmd", overwrite = TRUE)
 }