+++ R/dummy_imports.R
@@ -1,4 +1,3 @@
-#' @importFrom xaringan inf_mr
 #' @importFrom knitr kable
 #' @importFrom config get
 #' @importFrom fs file_exists