+++ R/build_slide.R
@@ -4,32 +4,6 @@ cross = clisymbols::symbol$cross
 tick = clisymbols::symbol$tick
 bullet = clisymbols::symbol$bullet
 
-check_graphics_paths = function() {
-  msg = glue::glue("{bullet} Checking graphics path...check_graphics_paths()")
-  message(yellow(msg))
-  if (!fs::file_exists("graphics")) {
-    fs::link_create("../graphics", "graphics", symbolic = TRUE)
-  }
-
-  has_error = FALSE
-  fnames = list.files(pattern = "^chapter?.\\.Rmd$")
-  for (fname in fnames) {
-    out = readLines(fname)
-    detect = stringr::str_detect(out, "../graphics/")
-    if (sum(detect) > 0L) {
-      lines = paste(which(detect), collapse = ", ")
-      if (!has_error) message(red("Change '../graphics/ to graphics in paths. See lines: "))
-      msg = glue::glue_col("  {cross} {lines} in file {fname}")
-      message(red(msg))
-      has_error = TRUE
-    }
-  }
-  if (has_error) stop(call. = FALSE)
-  msg = glue::glue("{tick} Graphics paths look good!")
-  message(yellow(msg))
-  return(invisible(NULL))
-}
-
 #' Convert RMD to HTML
 #'
 #' If the slides are not runtime shiny, the filename is passed to
@@ -39,6 +13,7 @@ check_graphics_paths = function() {
 #' @param fname Filename
 #' @export
 build_slide = function(fname) {
+  check_slides_title()
   check_graphics_paths()