+++ NEWS.md
@@ -1,6 +1,7 @@
 # Dev
   * Create symlink to graphics path
   * Check that titles in notes and slides match
+  * Use make final
 
 # Version 0.1.20
   * Add countdown suggests